Cleaning up utils.c
